Issue #1608: use double for docker.cpus property
Docker config option `--cpus` assumes float number, while current DMP
`docker.cpus` assumes integer number, which is actually `NanoCpus`.

Change this property to `double` and convert it to `NanoCpus` before
passing to Docker.

+++ src/main/java/io/fabric8/maven/docker/config/RunImageConfiguration.java
@@ -123,7 +123,7 @@ public class RunImageConfiguration implements Serializable {
     private Long cpuShares;
 
     @Parameter
-    private Long cpus;
+    private Double cpus;
 
     @Parameter
     private String cpuSet;
@@ -282,7 +282,7 @@ public class RunImageConfiguration implements Serializable {
         return cpuShares;
     }
 
-    public Long getCpus() {
+    public Double getCpus() {
         return cpus;
     }
 
@@ -632,7 +632,7 @@ public class RunImageConfiguration implements Serializable {
             return this;
         }
 
-        public Builder cpus(Long cpus) {
+        public Builder cpus(Double cpus) {
             config.cpus = cpus;
             return this;
         }
